% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellanneal.R
\name{run_cellanneal}
\alias{run_cellanneal}
\title{Wrapper for the cellanneals make_gene_dictionary and deconvolve functions}
\usage{
run_cellanneal(
  mixture,
  signature,
  sig_index = "gene",
  mix_index = "gene",
  disp_min = 0.5,
  bulk_min = 1e-05,
  bulk_max = 0.01,
  maxiter
)
}
\arguments{
\item{mixture}{A dataframe containing one column with gene names and several samples. Gene names must be unique.}

\item{signature}{A dataframe containg one columns with gene names and one for each cell type. Gene names must be unique.}

\item{sig_index}{Name of the gene name column in signature}

\item{mix_index}{Name of the gene name column in mixture}

\item{disp_min}{Minimum scaled dispersion. Cellanneal parameter}

\item{bulk_min}{Minimum allowed expression in mixture. Cellanneal parameter}

\item{bulk_max}{Maximum allowed expression in mixture. Cellanneal parameter}

\item{maxiter}{Maximum nbumber of iterations fpr dual annealing. Cellanneal parameter}
}
\value{
A dataframe containing cell-type proportions, Spearmans, and Pearsons rho for each sample.
}
\description{
Takes a mixture (bulk with gene column and several samples) and a signature (sc with gene colunm and several celltypes) and estimates cell-fractions in mixutre
}
\examples{
\dontrun{
run_cellanneal(mix, sig, sig_index = "index", mix_index = "gene")
}
}
